+++ src/simulation/regions.h
@@ -18,12 +18,12 @@ int get_region_y(int region_id, Spec spec);
 /**
  * Denormalizes an x-coordinate wrt region.
  */
-int denorm_region_x(long double x, int region_id, Spec spec);
+long double denorm_region_x(long double x, int region_id, Spec spec);
 
 /**
  * Denormalizes an y-coordinate wrt region.
  */
-int denorm_region_y(long double y, int region_id, Spec spec);
+long double denorm_region_y(long double y, int region_id, Spec spec);
 
 /**
  * Returns the horizon distance between two region IDs,
